DERBY-6299: Improve code coverage of org.apache.derby.iapi.services.sanity
This patch was contributed by Ahsan Shamsudeen (ahsan dot competition at gmail dot com)

This patch removes the classes SanityManager and AssertFailure from the
org.apache.derby.iapi.services.sanity package, replacing all uses of those
classes with references to the same-named classes in the
org.apache.derby.shared.common.sanity package.

This change is a follow-on to DERBY-1115, which moved the body of these
classes to the new location, but left the old classes around as wrappers.

Removing the wrapper classes simplifies the code, but is not intended to have
any other effect. There are no known bugs fixed by these changes, although
as pointed out during review, there may be behavior changes as a result of
this change. For example, code which specifies

	if (re instanceof AssertFailure)

may have different behavior now, since it is testing for an instance of the
underlying type, not an instance of the wrapper type.

This change affects a large set of files, but the change itself is quite simple.

Each of the files has a diff roughly like:

-import org.apache.derby.iapi.services.sanity.SanityManager;
+import org.apache.derby.shared.common.sanity.SanityManager;
 
In addition to modifying the import statement in all such .java files, we are
also deleting the two files

D       java\engine\org\apache\derby\iapi\services\sanity\AssertFailure.java
D       java\engine\org\apache\derby\iapi\services\sanity\SanityManager.java

These two files, left over from DERBY-1115, are trivial wrappers around the
same-named classes in o.a.d.shared.common.sanity package.

So all we're doing, then, is following up on DERBY-1115 by fixing the import
statements and removing the two wrapper classes.

So it's a big change, but conceptually quite simple.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1514927 13f79535-47bb-0310-9956-ffa450edef68