DERBY-6307: Improve code coverage of org.apache.derby.iapi.services.info
This patch was contributed by Ahsan Shamsudeen (ahsan dot competition at gmail dot com)

This patch adds several tests which directly exercise the utility methods in
org.apache.derby.iapi.services.info.JVMInfo.

Note that the precise code paths that are tested by these new test cases will
depend on the configuration of the test environment, since the JVMInfo code
is highly sensitive to the user's JDK level and JDK vendor.

The only JVMInfo method not directly tested at this point is javaDump,
which is a bit delicate to directly test because exercising this method can
give the appearance that a fatal error has occurred, which might confuse
somebody running the tests into thinking there was a real crash instead of
just an artificial execution of javaDump. So testing that method is deferred
until a later date.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1517149 13f79535-47bb-0310-9956-ffa450edef68