DERBY-5866( testFiringConstraintOrder(org.apache.derbyTesting.functionTests.tests.lang.TriggerTest)junit.framework.AssertionFailedError: matching triggers need to be fired in order creation:1,NO CASCADE BEFORE,DELETE,ROW)
As per Knut's recommendation, moving cleanup of TRIGGER_INFO back to tearDown(). Otherwise, the info from the last test case in TriggerTest won't be released until the JVM terminates. I don't expect it to take up a huge amount of memory, but with the big number of tests in suites.All unreleased data structures could accumulate and cause OOME.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1523821 13f79535-47bb-0310-9956-ffa450edef68