DERBY-534: Support use of the WHEN clause in CREATE TRIGGER statements
Add the WHEN clause syntax to the grammar and wire it together with the
existing partial code for the WHEN clause.

Make RowTriggerExecutor and StatementTriggerExecutor execute the WHEN
clause and use the result to decide whether the trigger action should
be executed.

Add some basic positive tests for the currently supported subset of the
functionality.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1523965 13f79535-47bb-0310-9956-ffa450edef68