DERBY-6320 Log a page dump to derby.log if ERROR nospc: nospc.U is returned to the user
backported change #1535075 from trunk to 10.10 branch, muliple manual changes
were necessary for the backport.

The backport of this change from trunk to 10.10 did not go cleanly at all due
to usage of new java language features in messages in trunk. Here is the patch
for 10.10 which I assume will backport cleanly to previous releases.
The 2 problems were a new 15 param message and that trunk did not require
Objects in newException() anymore. Also had to hand resolve a merge issue in
StoredPage.java

This patch adds the ability to dump a page in an insane build, and adds 2 calls
to do so in 2 outstanding nospc error cases. In those two cases a new user
level error is thrown and nests the nospc.U error so that we still know the
original stack trace where the lowest error was thrown.
The patch passes all tests and the specific errors were hand tested, one of
them using the test case filed in DERBY-4923 and in the other case just by hand
forcing the codepath.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.10@1535206 13f79535-47bb-0310-9956-ffa450edef68