DERBY-2041: Trigger should register a dependency on tables and columns used in its body
Add an upgrade test to verify how the dependencies are handled after
upgrade and downgrade.

Skip some cross checking between getColumns() and ResultSetMetaData
for views in DatabaseMetaDataTest, since those checks only pass for
tables. The upgrade test case adds a view that upsets the meta-data
test. (This is the first upgrade test that uses views.)

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1548032 13f79535-47bb-0310-9956-ffa450edef68