DERBY-6419 Make BTree scan honor OPENMODE_LOCK_NOWAIT for row locks
Patch derby-6419-1 introduces a mode to avoid waiting for row locks in
BTeee scans.  Using the debugger I can see that the current wait in
ConstraintCharactericsTest#testLocking, line 417 as of svn 1545394 is
no longer waiting with the patch, and stepping through the code in
IndexChanger I can see that the second next() which currently waits
returns immediately with an exception as desired.

Instead of overloading the current flag OPENMODE_LOCK_NOWAIT, I
introduced a new flag OPENMODE_LOCK_ROW_NOWAIT instead. The short
circuiting logic was added in B2IRowLocking3#LockRowOnPage: I had to
added an accessor to HeapController to get at the open_mode of the
base table's conglomerate; not sure if that's kosher or not. An
alternative could be to keep to flag in the B2IForwardScan object.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1550135 13f79535-47bb-0310-9956-ffa450edef68