DERBY-532 Support deferrable constraints
Patch derby-532-test-speedup changes ConstraintCharacteristicsTest to
use a main memory database for some tests for increased speed. It also
changes the way SystemPropertyTestSetup for static properties closes
down the database to not deregister the driver; without this change we
saw a test setup try to connect via the client driver to a Derby server
engine without a registered driver.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1550308 13f79535-47bb-0310-9956-ffa450edef68