DERBY-532 Support deferrable constraints
Patch derby-532-allow-pk-unique-1, which opens up for using deferrable
constraints for primary key and unique constraints, i.e. it is no
longer required that the special property "derby.constraintsTesting"
be used for those constraints, since the implementation is complete
modulo bugs. Upgrade tests still remain to be built, though.

For foreign key and check constraints as well as for "not enforced",
the property will still required till the implementation for those is
completed.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1555006 13f79535-47bb-0310-9956-ffa450edef68