DERBY-6443: ArrayIndexOutOfBoundsException when calling function from trigger
StaticMethodCallNode.bindExpression() was a no-op if the node had
already been bound. This caused problems for queries that need to bind
expressions multiple times (for example INSERT INTO ... SELECT).

This fix makes StaticMethodCallNode.bindExpression() work if it's
called a second time, while still making it a no-op if it is called
recursively in order to prevent infinite recursion.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1556809 13f79535-47bb-0310-9956-ffa450edef68