DERBY-4805(Increase the length of the RDBNAM field in the DRDA implementation)
This commit will allow RDBNAM to be 1024 bytes, thus lifting the limit from current 255bytes. I have changed the existing boundary test case to test for the new limit but the tests now use in-memory db jdbc url since on disk long urls were running into problem because of OS/file system limitation on the length of file name length.

As for newer client(version 10.11 and higher) going against an older server(10.10 and lower which will not have the fix for DERBY-4805 and hence will not accept the longer RDBNAM) with new RDBNAM limit, it will result in protocol exception. This is because when the client makes the first connect request to server, it has no way of knowing what server version it is talking to and hence there is no way to catch the length violation on the client side at this point.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1565491 13f79535-47bb-0310-9956-ffa450edef68