DERBY-6387: Incorrect ordering of triggers in dblook
backported change #1551940.  backport required some manual merge as the
key query had changed in multiple ways on the trunk.  Fix was to just add
the "order by" by hand to the existing 10.10 query.

Order the triggers by the CREATIONTIMESTAMP column so that dblook
produces a script that recreates the triggers in the same order as
they were originally created.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.10@1572994 13f79535-47bb-0310-9956-ffa450edef68