DERBY-5615: Permission problems with classpath subsubprotocol
Change the structure of CPFile.getInputStream() back to what it was
before the original fix for this issue (revision 1582655), but with
doPrivileged() calls around all operations that require privileges.

The restructuring in the original fix apparently prevented some
resources from being freed, so that DatabaseClassLoadingTest and
NativeAuthenticationServiceTest failed on Windows platforms because
they could not delete the jar file that contained the classpath
database.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1582754 13f79535-47bb-0310-9956-ffa450edef68