DERBY-6337(derbynet.ServerPropertiesTest.ttestSetPortPriority prints exception java.lang.Exception: DRDA_InvalidReplyTooShort.S:Invalid reply f rom network server: Insufficient data. but test passes)
Basically the issue is that sometimes when server is shutting down and a ping is sent to it around the same time, ping might be able to connect to server but before the ping hand shake can complete, the server might shutdown thus resulting in the ping client receiving a DRDA_InvalidReplyTooShort. This is probably why the test in this issue fails intermittently. To fix the issue, if ping does run into DRDA_InvalidReplyTooShort.S:Invalidreply, we will send another ping after waiting for a little time, and this time around we should get expected can't connect to server exception because server is down by now.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1594435 13f79535-47bb-0310-9956-ffa450edef68