DERBY-6576 A immediate Fk constraint blows up iff its referenced PK is deferred and we modify a duplicate key column
Patch derby-6576-forbid-deferred+cascade-set-null-1. It simplifies the
code of the previous patch for thi sissue that handles RESTRICT and NO
ACTION in combination with a deferrable primary key. It also forbids
CASCADE and SET NULL of the target primary key is deferrable, at least
for now until a satisfactory implementation can be found. Test have
been added for the new cases.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1599383 13f79535-47bb-0310-9956-ffa450edef68