DERBY-6552 The public api includes methods inherited from superclasses which aren't in the public api and so have no javadoc comments
Follow-up patch derby-6552-restructure-even-more. This restructures
the embedded hierarchy to avoid duplication also. It eliminates
EmbeddedBaseDataSource. Removed a slew of warnings in the IDE also by
some code washing, and regenerated the 10.11 serialized data sources
for the SerializedDataSourceTest. I removed some Javadoc from
non-public members in the API classes, and replaced them with
end-of-line comments instead (to avoid user confusion).

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1603022 13f79535-47bb-0310-9956-ffa450edef68