DERBY-6666 Deferred constraint validation fails with "dead statement" when query plan logging is enabled
Patch derby-6666c.

It removes this issue by adding push and push of the statement context
to the lcc before we execute the internal query to check for check
constraint violation. We had neglected to do this, which caused a
problem for the logging since the statement context should always be
properly initialized during execution.

This patch also adds a new test, derby6666 to
ConstraintCharacteristicsTest which fails without the fix in this
patch (equivalent to repro).

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1614796 13f79535-47bb-0310-9956-ffa450edef68