DERBY-6697: Failures when running tests with JaCoCo
Add permissions needed by JaCoCo to MissingPermissionTest's custom
policy file.

Make BaseTestCase.execJavaCmd() propagate the empty jacoco.active
property to sub-processes if the tests are running with JaCoCo.

Adjust dummy arguments given to procedures in DBOAccessTest so that
they fail with an error that is not affected by the broader
permissions granted when running with JaCoCo.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1617240 13f79535-47bb-0310-9956-ffa450edef68