DERBY-6715: MissingPermissionsTest fails under JaCoCo
Add permissions required by JaCoCo to the custom policy used by the test.

Change how SecurityManagerSetup checks for JaCoCo. By using
Class.forName() instead of getURL() we avoid the need for calling the
privileged getProtectionDomain() method, and the URL of the JaCoCo jar
file is not needed.

(It is usually OK to call this method without any special permissions
granted when the SecurityManagerSetup class is initialized. This is
because the SecurityManagerSetup is usually initialized before a
SecurityManager has been installed. In this test, however, a
subprocess is started with -Djava.security.manager, so a
SecurityManager is already in place before the SecurityManagerSetup
class is initialized in the subprocess.)

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1620793 13f79535-47bb-0310-9956-ffa450edef68